+++ src/threadpool.h
@@ -8,6 +8,24 @@
 extern "C" {
 #endif
 
+API_LOCAL int ev_loop_unlink_threadpool(ev_loop_t* loop);
+
+/**
+ * @brief Submit task into thread pool
+ * @warning This function is NOT MT-Safe and must be called in the thread where
+ *   \p loop is running.
+ * @param[in] pool      Thread pool
+ * @param[in] loop      Which event loop to call \p done_cb
+ * @param[in] token     Work token
+ * @param[in] type      Work type
+ * @param[in] work_cb   Work callback
+ * @param[in] done_cb   Work done callback
+ * @return              #ev_errno_t
+ */
+API_LOCAL int ev_threadpool_submit(ev_threadpool_t* pool, ev_loop_t* loop,
+    ev_threadpool_work_t* token, ev_threadpool_work_type_t type,
+    ev_threadpool_work_cb work_cb, ev_threadpool_work_done_cb done_cb);
+
 /**
  * @brief Submit task to threadpool.
  * @param[in] loop      Event loop.